requirements.txt
matplotlib==3.9.2
numpy==2.1.1
pandas==2.2.3
ray-default==2.35.0
ray-tune==2.35.0
scikit-learn==1.5.2
xgboost==2.1.1
jupyterlab
pip